move the hooks into a separate file
